+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.11.0] - September 6, 2019:
+
+* Make it so `width=100%` doesn't throw error. Fixes [#118](https://github.com/Sub6Resources/flutter_html/issues/118).
+* You can now set width and/or height in `ImageProperties` to negative to ignore the `width` and/or `height` values from the html. Fixes [#97](https://github.com/Sub6Resources/flutter_html/issues/97)
+* The `img` `alt` property now renders correctly when the image fails to load and with the correct style. Fixes [#96](https://github.com/Sub6Resources/flutter_html/issues/96)
+* Add partial support for `sub` tag.
+
 ## [0.10.4] - June 22, 2019:
 
 * Add support for `customTextStyle` to block and specialty HTML elements.